Separate business model from DB model (#1)
Initially I had some doubts where is the border between "As easy as possible" and good quality so decided to start with common Match model for business layer and "in memory DB", but decided that this separation is core quality requirement and separated those.
It also makes this MVP implementation more extendable in the future.

New features:

Created MatchEntity and adjusted repository implementation to work with it
Created entity mapper with Mapstruct
Added tests for new components
Adjusted existing tests

---------

* separate business model from DB model

* add missing tests

---------